ProcessWatcher.getProcessCpuTime returns Long, mulitple compatibility improvements

+++ src/main/java/io/github/kamilszewc/resourcewatcher/core/CpuInfo.java
@@ -17,6 +17,6 @@ public class CpuInfo {
     private Integer numberOfCoresPerSocket;
     private Integer numberOfThreadsPerSocket;
     private Integer numberOfSockets;
-    private Integer frequency;
+    private Float frequency;
 
 }
